Change how we format warnings/errors
We now show the line as it appeared in the original output, except shortening the path. Only the leading file component of the line is a link and its tooltip show the full path (but not potential line and column info, as that should now be in the visible output).